requirements.txt
# Core Django
django==5.0.6
gunicorn==21.2.0
whitenoise==6.6.0

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Auth
django-allauth>=0.63.3,<66.0.0

# Forms & UI
django-crispy-forms>=2.1
crispy-tailwind>=1.0.1

# AI & LangChain
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.13
langchain-core==0.2.11
openai==1.35.3

# Embeddings & Vector Search
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4

# Web Scraping / Fetching
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2

# Utilities
python-dotenv==1.0.1
Pillow==10.4.0
django-ratelimit==4.1.0

# Ollama (local AI)
ollama==0.2.1
